requirements.txt
python-dotenv==1.1.0
cryptography==45.0.3
fastapi==0.121.2
sqlalchemy==2.0.41
aiomysql==0.2.0
motor==3.7.1
neo4j==5.28.1
aioredis==2.0.1
aiokafka==0.8.1
python-jose==3.5.0
sentry-sdk==2.44.0
passlib==1.7.4
python-multipart==0.0.20
faiss-gpu==1.8.0
aiosmtplib==5.0.0
openai==1.91.0
langchain-core==0.3.63
langchain-community==0.3.24
starlette==0.46.2
